training/requirements.txt
torch==2.0.1
torchvision==0.15.2
diffusers==0.19.3
transformers==4.31.0
accelerate==0.21.0
datasets==2.12.0
wandb==0.15.5
numpy==1.24.3
pillow==8.3.2
opencv-python-headless==4.8.0.74
pandas==1.5.3
scikit-learn==1.2.2
matplotlib==3.7.1
tqdm==4.65.0
python-dotenv==0.19.0 